:lipstick: feat: Adição de arquivos para separar estilizações
