+++ components/dieta/index.tsx
@@ -8,6 +8,7 @@ import {
   ScrollView,
 } from "react-native";
 import { IDietaFixa } from "../../interfaces/IDieta";
+import { styles } from "./styles";
 
 interface DietaProps {
   dieta: IDietaFixa;
@@ -135,113 +136,5 @@ const DietaItem: React.FC<DietaProps> = ({
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  card: {
-    flex: 1,
-    borderWidth: 1,
-    borderColor: "#ddd",
-    borderRadius: 8,
-    overflow: "hidden",
-    margin: 8,
-    backgroundColor: "#fff",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 2,
-  },
-  detailsContainer: {
-    padding: 8,
-  },
-  nome: {
-    fontSize: 16,
-    fontWeight: "bold",
-    marginBottom: 4,
-  },
-  info: {
-    fontSize: 14,
-    color: "#555",
-    marginBottom: 2,
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-  },
-  modalContent: {
-    width: "90%",
-    padding: 20,
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.3,
-    shadowRadius: 5,
-    elevation: 5,
-    maxHeight: "80%",
-  },
-  scrollViewContent: {
-    paddingBottom: 20,
-  },
-  minimodalContent: {},
-  modalNome: {
-    fontSize: 18,
-    fontWeight: "bold",
-    marginBottom: 10,
-  },
-  minimodalNome: {
-    fontSize: 15,
-    fontWeight: "bold",
-  },
-  modalInfo: {
-    fontSize: 16,
-    color: "#555",
-    marginBottom: 8,
-  },
-  minimodalInfo: {
-    fontSize: 13,
-    color: "#555",
-    marginBottom: 5,
-  },
-  closeButton: {
-    marginTop: 15,
-    backgroundColor: "#2d74da",
-    paddingVertical: 8,
-    borderRadius: 5,
-    alignItems: "center",
-    width: "100%",
-  },
-  closeButtonText: {
-    color: "#fff",
-    fontSize: 16,
-  },
-  buttonContainer: {
-    flexDirection: "column",
-    justifyContent: "space-between",
-    paddingHorizontal: 10,
-    paddingVertical: 1,
-    marginBottom: 10,
-    gap: 3,
-  },
-  editButton: {
-    marginTop: 5,
-    backgroundColor: "#2d74da",
-    paddingVertical: 8,
-    borderRadius: 5,
-    alignItems: "center",
-    flex: 0.5,
-  },
-  deleteButton: {
-    marginTop: 5,
-    backgroundColor: "#e13f2f",
-    paddingVertical: 8,
-    borderRadius: 5,
-    alignItems: "center",
-    flex: 0.5,
-  },
-});
 
 export default DietaItem;
