+++ components/grupos/index.tsx
@@ -13,6 +13,7 @@ import { IGrupo } from "../../interfaces/IDieta"; // Importe a interface correta
 import { IAlimentoDieta } from "../../interfaces/IDieta";
 import { Ionicons } from "@expo/vector-icons";
 import resultChecker from "../../utils/resultChecker";
+import { styles } from "./styles";
 
 interface GroupModalProps {
   visible: boolean;
@@ -34,7 +35,9 @@ const GroupModal: React.FC<GroupModalProps> = ({
   onUpdateGrupo,
 }) => {
   const [localGrupo, setLocalGrupo] = useState<IGrupo | null>(grupo); // Estado local do grupo
-  const [editingAlimentoId, setEditingAlimentoId] = useState<string | null>(null);
+  const [editingAlimentoId, setEditingAlimentoId] = useState<string | null>(
+    null
+  );
   const [newPortion, setNewPortion] = useState<string>("");
   const [newQuantity, setNewQuantity] = useState<string>("");
 
@@ -49,7 +52,6 @@ const GroupModal: React.FC<GroupModalProps> = ({
     }
   }, [visible]);
 
-
   // Função que atualiza o grupo local e envia para o pai
   const updateAndSendGrupo = (updatedAlimentos: IAlimentoDieta[]) => {
     if (localGrupo) {
@@ -61,8 +63,9 @@ const GroupModal: React.FC<GroupModalProps> = ({
 
   const handleEditQuantity = (alimentoId: string) => {
     if (!resultChecker.checkQuantidade(newQuantity)) {
-      return; // Retorna se a porção for inválida
+      return;
     }
+
     const quantity = Number(newQuantity);
     if (!isNaN(quantity) && quantity > 0) {
       onEditQuantity(alimentoId, quantity);
@@ -81,8 +84,9 @@ const GroupModal: React.FC<GroupModalProps> = ({
 
   const handleEditPortion = (alimentoId: string) => {
     if (!resultChecker.checkPorcao(newPortion)) {
-      return; // Retorna se a porção for inválida
+      return;
     }
+
     const portion = Number(newPortion);
     if (!isNaN(portion) && portion > 0) {
       onEditPortion(alimentoId, portion);
@@ -133,7 +137,6 @@ const GroupModal: React.FC<GroupModalProps> = ({
     }
   };
 
-
   if (!localGrupo) return null; // Se o grupo não estiver definido
 
   return (
@@ -150,15 +153,19 @@ const GroupModal: React.FC<GroupModalProps> = ({
                   onLongPress={() => handleLongPress(alimento.alimentoId)}
                   style={styles.alimentoContainer}
                 >
-                  
                   <Text style={styles.modalInfo}>
-                    {alimento.quantidade}x - {alimento.nome} - {alimento.porcao}g
+                    {alimento.quantidade}x - {alimento.nome} - {alimento.porcao}
+                    g
                   </Text>
-                  <Ionicons 
-                            name={editingAlimentoId === alimento.alimentoId ? "caret-up" : "caret-down"} 
-                            size={24} 
-                            color={"#000"} 
-                        />
+                  <Ionicons
+                    name={
+                      editingAlimentoId === alimento.alimentoId
+                        ? "caret-up"
+                        : "caret-down"
+                    }
+                    size={24}
+                    color={"#000"}
+                  />
                 </TouchableOpacity>
                 {editingAlimentoId === alimento.alimentoId && (
                   <>
@@ -206,82 +213,5 @@ const GroupModal: React.FC<GroupModalProps> = ({
   );
 };
 
-const styles = StyleSheet.create({
-  modalContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-  },
-  modalContent: {
-    width: "90%",
-    padding: 20,
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    maxHeight: "80%",
-  },
-  scrollViewContent: {
-    paddingBottom: 20,
-  },
-  modalNome: {
-    fontSize: 20,
-    fontWeight: "bold",
-    marginBottom: 10,
-    alignSelf: "center"
-  },
-  modalSubNome: {
-    fontSize: 18,
-    fontWeight: "bold",
-    marginBottom: 10,
-  },
-  modalInfo: {
-    fontSize: 16,
-    color: "#555",
-    marginBottom: 8,
-  },
-  closeButton: {
-    marginTop: 15,
-    backgroundColor: "#2d74da",
-    paddingVertical: 8,
-    borderRadius: 5,
-    alignItems: "center",
-    width: "100%",
-  },
-  closeButtonText: {
-    color: "#fff",
-    fontSize: 16,
-  },
-  alimentoContainer: {
-    paddingVertical: 10,
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    gap: 10
-  },
-  editContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginTop: 5,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: "#ddd",
-    borderRadius: 5,
-    padding: 5,
-    flex: 1,
-    marginRight: 5,
-  },
-  saveButton: {
-    backgroundColor: "#2d74da",
-    paddingVertical: 8,
-    paddingHorizontal: 10,
-    borderRadius: 5,
-  },
-  saveButtonText: {
-    color: "#fff",
-    fontSize: 16,
-  },
-});
 
 export default GroupModal;
