+++ screens/cadastro/styles.ts
@@ -1,10 +1,11 @@
 import { StyleSheet } from "react-native";
+import colors from "../../colors/colors";
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     flexDirection: "column",
-    backgroundColor: "#fff",
+    backgroundColor: colors.background,
     paddingHorizontal: 20,
     justifyContent: "center",
     alignItems: "center",
@@ -65,7 +66,7 @@ const styles = StyleSheet.create({
     color: "#007bff",
   },
   button: {
-    backgroundColor: "#007bff",
+    backgroundColor: colors.blueButtonCollor,
     paddingVertical: 18,
     paddingHorizontal: "30%",
     borderRadius: 30,
