+++ screens/home/styles.ts
@@ -1,10 +1,12 @@
 import { StyleSheet } from 'react-native';
+import colors from '../../colors/colors';
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     flexDirection: "column",
     justifyContent: 'center',
+    backgroundColor: colors.background
   },
   containercima: {
     flex: 0.5,
