+++ screens/login/styles.ts
@@ -1,5 +1,6 @@
 
 import { StyleSheet } from 'react-native';
+import colors from '../../colors/colors';
 
 const styles = StyleSheet.create({
     container: {
@@ -7,7 +8,8 @@ const styles = StyleSheet.create({
         flexDirection: "column",
         justifyContent: 'flex-start', // Alinha os itens no topo
         width: '100%',
-        paddingTop: 50 // Adiciona um pouco de espaçamento no topo
+        paddingTop: 50, // Adiciona um pouco de espaçamento no topo
+        backgroundColor: colors.background
     },
     titulo: {
         fontSize: 24,
