+++ screens/userAlimentos/styles.ts
@@ -1,11 +1,9 @@
 import { StyleSheet } from "react-native";
-import colors from "../../colors/colors";
 
-const styles = StyleSheet.create({
+export const styles = StyleSheet.create({
     container: {
         flex: 1,
         padding: 16,
-        backgroundColor: colors.background
     },
     title: {
         fontSize: 24,
@@ -15,7 +13,7 @@ const styles = StyleSheet.create({
     picker: {
         height: 50,
         width: '100%',
-        marginBottom: 10,
+        marginBottom: 1,
     },
     row: {
         justifyContent: 'space-between',
@@ -29,7 +27,7 @@ const styles = StyleSheet.create({
         color: '#007bff',
     },
     button: {
-        backgroundColor: colors.blueButtonCollor,
+        backgroundColor: "#007bff",
         paddingVertical: 9,
         paddingHorizontal: "25%",
         borderRadius: 30,
@@ -50,6 +48,21 @@ const styles = StyleSheet.create({
     icon: {
         marginRight: 10,
     },
+    searchContainer: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        backgroundColor: '#fff',
+        padding: 5,
+        borderRadius: 25,
+        marginTop: 10,
+        shadowColor: '#000',
+        shadowOpacity: 0.1,
+        shadowOffset: { width: 0, height: 5 },
+        elevation: 5,
+    },
+    searchInput: {
+        marginLeft: 10,
+        flex: 1,
+        fontSize: 16
+    },
 });
-
-export default styles
